debug - float to decimal for prices

+++ CostOfRevenue/Models/Furniture.cs
@@ -20,14 +20,14 @@ namespace CostOfRevenue.Models
         [ObservableProperty]
         private Enums.Unit _unit;
         [ObservableProperty]
-        private float _unitPrice;
+        private decimal _unitPrice;
         [ObservableProperty]
         private DateTime _date;
         [ObservableProperty]
         private float _stockQuantity;
 
         [JsonConstructor]
-        public Furniture(string id, string name, Enums.Unit unit, float unitPrice, DateTime date, float stockQuantity)
+        public Furniture(string id, string name, Enums.Unit unit, decimal unitPrice, DateTime date, float stockQuantity)
         {
             _id = id;
             _name = name;
@@ -36,7 +36,7 @@ namespace CostOfRevenue.Models
             _date = date.Date;
             _stockQuantity = stockQuantity;
         }
-        public Furniture(string id, string name, Enums.Unit unit, float unitPrice, DateTime date)
+        public Furniture(string id, string name, Enums.Unit unit, decimal unitPrice, DateTime date)
         {
             _id = id;
             _name = name;
